consolidated types defined in 3 places to common
We had several types defined in:

* backend/functions/lib/use-cases/case-docket/case-docket.model.ts
* user-interface/src/lib/type-declarations/chapter-15.d.ts
* user-interface/src/lib/type-declarations/api.d.ts
* common/

These have all been verified and consolided to common now.
Various modules and tests needed updating to use the types in common.

Jira ticket: CAMS-356
