--- backend/functions/lib/use-cases/case-docket/case-docket.model.ts
@@ -1,17 +1,0 @@
-export type CaseDocketEntryDocument = {
-  fileUri: string;
-  fileSize: number;
-  fileLabel: string;
-  fileExt?: string;
-};
-
-export type CaseDocketEntry = {
-  sequenceNumber: number;
-  documentNumber?: number;
-  dateFiled: string;
-  summaryText: string;
-  fullText: string;
-  documents?: CaseDocketEntryDocument[];
-};
-
-export type CaseDocket = Array<CaseDocketEntry>;
